src/backend/requirements.txt
numpy==1.13.1
urllib3==1.22
gensim==2.2.0
requests==2.18.4
pycrypto==2.6.1
argon2-cffi==18.1.0
colorlog==3.1.4
pymongo==3.6.1
scikit_learn==0.19.1
ujson==1.35
websockets==4.0.1
circus==0.14.0
tornado==4.5.2
maxminddb==1.3.0
maxminddb-geolite2==2017.803
